+++ packages/livedoc-mocha/docs/ReleaseNotes.md
@@ -1,4 +1,7 @@
 # Release Notes
+## 0.4.0
+* Public release
+
 ## 0.4.0-beta-7
 * Bug: [#70](https://github.com/dotnetprofessional/LiveDoc/issues/70) Errors occurring in afterBackground not displaying.